Initial stab at implementing Atomic based on GreeGeer's code
Initial stab at implementing Atomic based on GreeGeer's code
(as I remember it)

This commit adds an atomic block

atomic a <- b do
  a ! atomic()
  b ! non_atomic()
end

which is desugared into

let tmp = b in
  atomic a <- tmp do
    atomicStart(tmp)
    a ! atomic()
    b ! non_atomic()
    atomicStop(tmp)
  end

where a has type Atomic(T) if b has type T

Atomic(T) implies StackBound(T)

In the optimiser, atomic blocks are dropped and the code that gets
compiled is ultimately

let tmp = b in
  let a = tmp in
    atomicStart(tmp)
    a ! atomic()
    b ! non_atomic()
    atomicStop(tmp)
  end

In the CodeGen pass, there are now TWO call nodes:

Call (the old one)
AtomicCall (the new one)

In Expr.hs, (possibly incomplete -- not much testing!) we check the
target type of message sends and if we find a message send to a target
whose type is Atomic, we use AtomicCall, else Call

Finally, in the PrettyCode pass, we output AtomicCall and Call nodes
exactly the same way, except that the AtomicCall inserts a ATOMIC
comment in the C source.

TODOs:

1.
Edit Std.enc and change functions startAtomic and stopAtomic (by
dropping to C level withing Encore) to send the start and stop atomic
messages. There is quite likely a better way to do this, but this felt
like an OK start just to get something going.

2.
Change the code generation so that the C function that get generated
for each possible message send takes an additional boolean parameter
"atomic" and if that is set to true, set the corresponding flag in
the messages sent to the actor.